+++ src/Sentry/Laravel/Integration.php
@@ -149,7 +149,14 @@ class Integration implements IntegrationInterface
 
         if (empty($routeName) && $route->getActionName()) {
             // SomeController@someAction (controller action)
-            $routeName = ltrim($route->getActionName(), (self::$baseControllerNamespace ?? '') . '\\');
+            $routeName = $route->getActionName();
+
+            $baseNamespace = self::$baseControllerNamespace ?? '';
+
+            // Strip away the base namespace from the action name
+            if (!empty($baseNamespace)) {
+                $routeName = Str::after($routeName, $baseNamespace . '\\');
+            }
         }
 
         if (empty($routeName) || $routeName === 'Closure') {