+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix incorrectly stripped base controller action from transaction name (#406)
+
 ## 2.1.1
 
 - Fix for potential `Undefined index: controllers_base_namespace.` notice